hotfix: have main page of the docker docs at the first place on the tags page
currently the pages on this site can only be accessed through the 'tags' menu, where pages are listed under tags, ordered by their name.
since (for now) this is the only place where the docs are available, having the starting page be at the first place is useful for directing readers to the correct page.

this should only be a temporary fix, until a solution is found to display starting pages on the website